Using Columns names instead of ORM to get all documents (#620)
* Using Columns name instead of ORM object for get all documents call

* Separating meta search from documents. This way it will optimize the memory not duplicating document.text

* Fixing mypy issue

* SQLite have limit on number of host variable hence using batching to fetch meta information

* Query meta only if meta field is not Null in DocOrm

* Add batch_size to other functions except label

* meta can be none so fix that issue

* Dummy commit to trigger CI

* Using chunked dictionary

* Upgrading faiss

* reverting change related to  faiss upgrade

* Changing DB name in test_faiss_retrieving test as it might interfere with exiting files by corrupting DB file

* Updating doc string related to batch_size

* Update docstring for batch_size

+++ haystack/document_store/sql.py
@@ -1,3 +1,4 @@
+import itertools
 import logging
 from typing import Any, Dict, Union, List, Optional
 from uuid import uuid4
@@ -42,7 +43,12 @@ class MetaORM(ORMBase):
 
     name = Column(String(100), index=True)
     value = Column(String(1000), index=True)
-    document_id = Column(String(100), ForeignKey("document.id", ondelete="CASCADE", onupdate="CASCADE"), nullable=False)
+    document_id = Column(
+        String(100),
+        ForeignKey("document.id", ondelete="CASCADE", onupdate="CASCADE"),
+        nullable=False,
+        index=True
+    )
 
     documents = relationship(DocumentORM, backref="Meta")
 
@@ -69,6 +75,7 @@ class SQLDocumentStore(BaseDocumentStore):
         index: str = "document",
         label_index: str = "label",
         update_existing_documents: bool = False,
+        batch_size: int = 32766,
     ):
         """
         An SQL backed DocumentStore. Currently supports SQLite, PostgreSQL and MySQL backends.
@@ -80,7 +87,13 @@ class SQLDocumentStore(BaseDocumentStore):
         :param update_existing_documents: Whether to update any existing documents with the same ID when adding
                                           documents. When set as True, any document with an existing ID gets updated.
                                           If set to False, an error is raised if the document ID of the document being
-                                          added already exists. Using this parameter coud cause performance degradation for document insertion. 
+                                          added already exists. Using this parameter could cause performance degradation
+                                          for document insertion.
+        :param batch_size: Maximum number of variable parameters and rows fetched in a single SQL statement,
+                           to help in excessive memory allocations. In most methods of the DocumentStore this means number of documents fetched in one query.
+                           Tune this value based on host machine main memory.
+                           For SQLite versions prior to v3.32.0 keep this value less than 1000.
+                           More info refer: https://www.sqlite.org/limits.html
         """
         engine = create_engine(url)
         ORMBase.metadata.create_all(engine)
@@ -91,6 +104,7 @@ class SQLDocumentStore(BaseDocumentStore):
         self.update_existing_documents = update_existing_documents
         if getattr(self, "similarity", None) is None:
             self.similarity = None
+        self.batch_size = batch_size
 
     def get_document_by_id(self, id: str, index: Optional[str] = None) -> Optional[Document]:
         """Fetch a document by specifying its text id string"""
@@ -101,21 +115,33 @@ class SQLDocumentStore(BaseDocumentStore):
     def get_documents_by_id(self, ids: List[str], index: Optional[str] = None) -> List[Document]:
         """Fetch documents by specifying a list of text id strings"""
         index = index or self.index
-        results = self.session.query(DocumentORM).filter(DocumentORM.id.in_(ids), DocumentORM.index == index).all()
-        documents = [self._convert_sql_row_to_document(row) for row in results]
+
+        documents = []
+        for i in range(0, len(ids), self.batch_size):
+            query = self.session.query(DocumentORM).filter(
+                DocumentORM.id.in_(ids[i: i + self.batch_size]),
+                DocumentORM.index == index
+            )
+            for row in query.all():
+                documents.append(self._convert_sql_row_to_document(row))
 
         return documents
 
     def get_documents_by_vector_ids(self, vector_ids: List[str], index: Optional[str] = None):
         """Fetch documents by specifying a list of text vector id strings"""
         index = index or self.index
-        results = self.session.query(DocumentORM).filter(
-            DocumentORM.vector_id.in_(vector_ids),
-            DocumentORM.index == index
-        ).all()
-        sorted_results = sorted(results, key=lambda doc: vector_ids.index(doc.vector_id))
-        documents = [self._convert_sql_row_to_document(row) for row in sorted_results]
-        return documents
+
+        documents = []
+        for i in range(0, len(vector_ids), self.batch_size):
+            query = self.session.query(DocumentORM).filter(
+                DocumentORM.vector_id.in_(vector_ids[i: i + self.batch_size]),
+                DocumentORM.index == index
+            )
+            for row in query.all():
+                documents.append(self._convert_sql_row_to_document(row))
+
+        sorted_documents = sorted(documents, key=lambda doc: vector_ids.index(doc.meta["vector_id"]))  # type: ignore
+        return sorted_documents
 
     def get_all_documents(
             self,
@@ -134,21 +160,52 @@ class SQLDocumentStore(BaseDocumentStore):
         """
 
         index = index or self.index
-        query = self.session.query(DocumentORM).filter_by(index=index)
+        # Generally ORM objects kept in memory cause performance issue
+        # Hence using directly column name improve memory and performance.
+        # Refer https://stackoverflow.com/questions/23185319/why-is-loading-sqlalchemy-objects-via-the-orm-5-8x-slower-than-rows-via-a-raw-my
+        documents_query = self.session.query(
+            DocumentORM.id,
+            DocumentORM.text,
+            DocumentORM.vector_id
+        ).filter_by(index=index)
 
         if filters:
-            query = query.join(MetaORM)
+            documents_query = documents_query.join(MetaORM)
             for key, values in filters.items():
-                query = query.filter(MetaORM.name == key, MetaORM.value.in_(values))
+                documents_query = documents_query.filter(
+                    MetaORM.name == key,
+                    MetaORM.value.in_(values),
+                    DocumentORM.id == MetaORM.document_id
+                )
 
-        documents = [self._convert_sql_row_to_document(row) for row in query.all()]
-        return documents
+        documents_map = {}
+        for row in documents_query.all():
+            documents_map[row.id] = Document(
+                id=row.id,
+                text=row.text,
+                meta=None if row.vector_id is None else {"vector_id": row.vector_id} # type: ignore
+            )
+
+        for doc_ids in self.chunked_iterable(documents_map.keys(), size=self.batch_size):
+            meta_query = self.session.query(
+                MetaORM.document_id,
+                MetaORM.name,
+                MetaORM.value
+            ).filter(MetaORM.document_id.in_(doc_ids))
+
+            for row in meta_query.all():
+                if documents_map[row.document_id].meta is None:
+                    documents_map[row.document_id].meta = {}
+                documents_map[row.document_id].meta[row.name] = row.value # type: ignore
+
+        return list(documents_map.values())
 
     def get_all_labels(self, index=None, filters: Optional[dict] = None):
         """
         Return all labels in the document store
         """
         index = index or self.label_index
+        # TODO: Use batch_size
         label_rows = self.session.query(LabelORM).filter_by(index=index).all()
         labels = [self._convert_sql_row_to_label(row) for row in label_rows]
 
@@ -169,33 +226,41 @@ class SQLDocumentStore(BaseDocumentStore):
         :return: None
         """
 
-        # Make sure we comply to Document class format
-        document_objects = [Document.from_dict(d) if isinstance(d, dict) else d for d in documents]
         index = index or self.index
-        for doc in document_objects:
-            meta_fields = doc.meta or {}
-            vector_id = meta_fields.get("vector_id")
-            meta_orms = [MetaORM(name=key, value=value) for key, value in meta_fields.items()]
-            doc_orm = DocumentORM(id=doc.id, text=doc.text, vector_id=vector_id, meta=meta_orms, index=index)
-            if self.update_existing_documents:
-                # First old meta data cleaning is required
-                self.session.query(MetaORM).filter_by(document_id=doc.id).delete()
-                self.session.merge(doc_orm)
-            else:
-                self.session.add(doc_orm)
-        try:
-            self.session.commit()
-        except Exception as ex:
-            logger.error(f"Transaction rollback: {ex.__cause__}")
-            # Rollback is important here otherwise self.session will be in inconsistent state and next call will fail
-            self.session.rollback()
-            raise ex
+        if len(documents) == 0:
+            return
+        # Make sure we comply to Document class format
+        if isinstance(documents[0], dict):
+            document_objects = [Document.from_dict(d) if isinstance(d, dict) else d for d in documents]
+        else:
+            document_objects = documents
+
+        for i in range(0, len(document_objects), self.batch_size):
+            for doc in document_objects[i: i + self.batch_size]:
+                meta_fields = doc.meta or {}
+                vector_id = meta_fields.get("vector_id")
+                meta_orms = [MetaORM(name=key, value=value) for key, value in meta_fields.items()]
+                doc_orm = DocumentORM(id=doc.id, text=doc.text, vector_id=vector_id, meta=meta_orms, index=index)
+                if self.update_existing_documents:
+                    # First old meta data cleaning is required
+                    self.session.query(MetaORM).filter_by(document_id=doc.id).delete()
+                    self.session.merge(doc_orm)
+                else:
+                    self.session.add(doc_orm)
+            try:
+                self.session.commit()
+            except Exception as ex:
+                logger.error(f"Transaction rollback: {ex.__cause__}")
+                # Rollback is important here otherwise self.session will be in inconsistent state and next call will fail
+                self.session.rollback()
+                raise ex
 
     def write_labels(self, labels, index=None):
         """Write annotation labels into document store."""
 
         labels = [Label.from_dict(l) if isinstance(l, dict) else l for l in labels]
         index = index or self.label_index
+        # TODO: Use batch_size
         for label in labels:
             label_orm = LabelORM(
                 document_id=label.document_id,
@@ -220,16 +285,22 @@ class SQLDocumentStore(BaseDocumentStore):
         :param index: filter documents by the optional index attribute for documents in database.
         """
         index = index or self.index
-        self.session.query(DocumentORM).filter(
-            DocumentORM.id.in_(vector_id_map),
-            DocumentORM.index == index
-        ).update({
-            DocumentORM.vector_id: case(
-                vector_id_map,
-                value=DocumentORM.id,
-            )
-        }, synchronize_session=False)
-        self.session.commit()
+        for chunk_map in self.chunked_dict(vector_id_map, size=self.batch_size):
+            self.session.query(DocumentORM).filter(
+                DocumentORM.id.in_(chunk_map),
+                DocumentORM.index == index
+            ).update({
+                DocumentORM.vector_id: case(
+                    chunk_map,
+                    value=DocumentORM.id,
+                )
+            }, synchronize_session=False)
+            try:
+                self.session.commit()
+            except Exception as ex:
+                logger.error(f"Transaction rollback: {ex.__cause__}")
+                self.session.rollback()
+                raise ex
 
     def update_document_meta(self, id: str, meta: Dict[str, str]):
         """
@@ -338,3 +409,17 @@ class SQLDocumentStore(BaseDocumentStore):
             session.add(instance)
             session.commit()
             return instance
+
+    # Refer: https://alexwlchan.net/2018/12/iterating-in-fixed-size-chunks/
+    def chunked_iterable(self, iterable, size):
+        it = iter(iterable)
+        while True:
+            chunk = tuple(itertools.islice(it, size))
+            if not chunk:
+                break
+            yield chunk
+
+    def chunked_dict(self, dictionary, size):
+        it = iter(dictionary)
+        for i in range(0, len(dictionary), size):
+            yield {k: dictionary[k] for k in itertools.islice(it, size)}
